Author: Dominion Aromolaran - psyda5 Date: 17th December. Commit Type: Chore
Enhancing MusicPlayer

Removed unnecessary manual threading for the
MediaPlayer as it handles its own threading.

Implemented a stop method that stops playback and
releases resources with the dispose() method.

Ensured fields are private and cannot be modified
from outside the class.

Added a method setMusic to change the music track if
needed.